refactor: move credentials loader to internal/sourcesync (#807)
LoadGitCredentials and LoadOCICredentials move from
ee/internal/controller to internal/sourcesync. They work against the
(already-core) GitCredentials / OCICredentials types and are needed
by the upcoming core SkillSource controller.

Arena controllers call the package-qualified forms
(sourcesync.LoadGitCredentials, sourcesync.LoadOCICredentials). Test
file moved verbatim modulo package/license header.

No behaviour change.

Ref #807
